% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_function.R
\name{calibrate_collinearity}
\alias{calibrate_collinearity}
\title{Calibrate Collinearity for Simulation (Target VIF Values)}
\usage{
calibrate_collinearity()
}
\description{
Provides utilities for mapping user-defined collinearity levels to
approximate VIF ranges used in simulation design.
}
