% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{rdata_fun}
\alias{rdata_fun}
\title{Data Generating Process}
\usage{
rdata_fun(x, f, eps)
}
\arguments{
\item{x}{either a numeric vector, a numeric matrix or a function. When called
on a scalar integer argument must return a vector of the same length.
See Details.}

\item{f, eps}{functions. When called on a single numeric vector or matrix must
return a numeric vector with the same number of observations (as given by
\code{NROW()}, \emph{cf.} \link[base]{nrow}). See Details.}
}
\value{
A function of no arguments in the fixed \eqn{X} case,
and of an integer argument \code{n} in the random \eqn{X} case. See Details.
}
\description{
Generate datasets \eqn{\{(X_i, Y_i)\}_{i = 1, 2, \dots, n}} paired
observations according to a specified data generating process:

\deqn{Y_i = f(X_i) + \varepsilon _i.}

The building blocks of the data generating process, \emph{i.e.} the function
\eqn{f} and the distributions of \eqn{X_i} and \eqn{\varepsilon _i} are
provided as arguments. Can handle both fixed and random regressors \eqn{X_i}.
}
\details{
The only purpose of this function is to provide a convenience wrapper for
generating \eqn{XY} datasets according to the scheme described in the
description. \code{rdata_fun()} returns a function that can be called for this
purpose.
\subsection{Conditional mean and conditional noise}{

The interpretation of the \eqn{f} and \eqn{\varepsilon} terms in the equation
above is the following:
\itemize{
\item \eqn{f(X_i) = \mathbb E(Y_i \vert X_i)}. This is the conditional
expectation of the responses \eqn{Y_i}, given the value of the regressors
\eqn{X_i}.
\item \eqn{\varepsilon _i = Y_i\vert X_i - \mathbb E(Y_i \vert X_i)}. The noise
of the response, that may in principle depend on \eqn{X_i}.
}

These terms are encoded, respectively, by the \code{f} and \code{eps} arguments of
\code{rdata_fun()}. These functions are supposed to be called with a single
numeric vector or matrix argument, and return a vector with the same number
of observations, as reported by \code{NROW()}.

It is understood that the outputs of \code{f} and \code{eps} should be
deterministic and random, respectively, and that \code{eps} should have zero
conditional (on its argument) mean.
\emph{No automatic check is performed to ensure these expectations are satisfied},
and the correct definition of these argument is ultimately left to the user.
See examples below.
}

\subsection{Fixed vs. Random regressors}{

As for the regressors, both cases of fixed and random \eqn{X_i} are
contemplated:
\itemize{
\item For \emph{fixed} \eqn{X_i}'s, \code{x} should be specified as a numeric vector or
matrix, corresponding to the fixed values for the regressors.
\item For \emph{random} \eqn{X_i}'s, \code{x} should be specified as a function of a single
integer argument, a random generator for \eqn{X_i} observations (the argument
corresponding to the variable number of observations to draw from \eqn{X_i}
distribution).
}
}

\subsection{Returned value}{

The returned value is a function of no arguments in the fixed \eqn{X} case,
and of an integer argument \code{n} in the random \eqn{X} case. Calls to this
function return a \code{data.frame} of observations drawn from the specified
\eqn{\{(X_i, Y_i)\}_{i = 1, 2, \dots, n}} distribution.
}
}
\examples{
# Linear model: Y = q + m X + eps
rxy <- rdata_fun(
  x = rnorm,
  f = \(x) 1 + 2*x,
  eps = \(x) rnorm(length(x), sd = 0.1)
)

# calling:
rxy(10)
# does the same job as:
x <- rnorm(10)
y <- 1 + 2 * x + rnorm(10, sd = 0.1)
data.frame(x, y)

}
