% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_function.R
\name{run_diagnostics}
\alias{run_diagnostics}
\title{Linear Model Diagnostics (Safe Shapiro Test)}
\usage{
run_diagnostics(fit, data_obs)
}
\arguments{
\item{fit}{LS/GLM fit}

\item{data_obs}{Observed data}
}
\value{
List of diagnostics (VIF, Cook's distance, residual plots)
}
\description{
Produces VIF, Cook’s distance, residual diagnostics, and a
safe Shapiro-Wilk normality test that automatically handles
sample-size limitations.
}
